chore: reorganize folder structure
