+++ frontend/src/components/app.tsx
@@ -5,10 +5,10 @@ import { Header } from './header/header';
 import { SearchBox } from './search-box/search-box';
 import { ResultList } from './result-list/result-list';
 import { UserIcon, SkillIcon } from './svg-icons/svg-icons';
-import { LoginPage } from './login-page';
-import { MyProfilePage } from './my-profile-page';
-import { ProfilePage } from './profile-page';
-import { SkillPage } from './skill-page';
+import { LoginPage } from './login-page/login-page';
+import { MyProfilePage } from './my-profile-page/my-profile-page';
+import { ProfilePage } from './profile-page/profile-page';
+import { SkillPage } from './skill-page/skill-page';
 import { SkillApi, Identity, UserSkill } from '../api/skill-api';
 import { ApiException } from '../api/ajax';
 import { Toaster } from './toaster/toaster';
