+++ frontend/src/components/error-list/error-list.scss
@@ -1,0 +1,18 @@
+.error-list {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+
+  &__item {
+    padding: 8px;
+    margin: 8px 0;
+    background: #7f003f;
+    color: #fff;
+    border-radius: 5px;
+    border: 1px solid #ff007f;
+
+    &:first-child {
+      margin: 0;
+    }
+  }
+}
