+++ frontend/src/components/text-input/text-input.tsx
@@ -8,17 +8,18 @@ type TextInputProps = {
   type?: string;
   onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
   value: any;
+  required ?: boolean;
   disabled ?: boolean;
   readOnly ?: boolean;
 }
 
 
 export const TextInput = React.forwardRef<HTMLInputElement, TextInputProps>(({
-  name, id, type = 'text', value, onChange, className = '', disabled = false, readOnly = false
+  name, id, type = 'text', value, onChange, className = '', disabled = false, readOnly = false, required = false
 }, ref) => {
   return (
     <input className={'text-input ' + className}
       name={name} ref={ref} id={id} type={type} 
-      onChange={onChange} value={value} disabled={disabled} readOnly={readOnly} />
+      onChange={onChange} value={value} disabled={disabled} readOnly={readOnly} required={required} />
   );
 });
