+++ frontend/src/css/index.scss
@@ -1,6 +1,5 @@
 @import './vars';
 @import './base';
 @import './forms';
-@import './login';
 @import './profile-page';
 @import './skill-page';
